update error handling for API requests
